+++ src/ui/Forms/Styles/SubStationAlphaStylesBatchConvert.cs
@@ -831,9 +831,9 @@ namespace Nikse.SubtitleEdit.Forms.Styles
             checkBoxFontBold.Checked = style.Bold;
             checkBoxFontUnderline.Checked = style.Underline;
 
-            if (style.FontSize > 0 && style.FontSize <= (float)numericUpDownFontSize.Maximum)
+            if (style.FontSize > 0 && style.FontSize <= numericUpDownFontSize.Maximum)
             {
-                numericUpDownFontSize.Value = (decimal)style.FontSize;
+                numericUpDownFontSize.Value = style.FontSize;
             }
             else
             {
